Move report to app layer
